+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Unreleased in the current development version (target v0.24.0):
 
 Complete list:
 - Fallback test download from wilma (#269)
-- Jinja templates for configuration files for collections (#230)
+- Jinja templates for configuration files for collections (#230, #270)
 - Histogram: diagnostic updates (#255)
 - Seaice: Port changes from Dashboard v0.19-op  (#190)
 - ECmean: prevent creation of default empty dirs (#263)